third-party: Disable sh detection with mingw-make
When building with MinGW Makefiles, cmake refuses to generate a project
if sh.exe is found in path.

It would seem this is no longer a problem with Make

    https://cmake.org/pipermail/cmake/2016-June/063739.html

but the cmake check remains in place and we need to work around it, in
order to build inside the Mingw console.